progress1222

+++ src/routes/evaluate.js
@@ -27,67 +27,70 @@ let connection = mysql.createConnection({
 const query = util.promisify(connection.query).bind(connection);
 router.get('/', function(req, res, next){
   let profession;
-  let bool;
+  let bool = false;
   let obj;
   let sender;
   let receiver;
   let eventlist;
   async function getESR(){
     const users = await query('select * from users;');
-    // if (req.user.email) {
-    //   for (i = 0; i < users.length; i++) {
-    //     if(req.user.email == users[i].email){
-    //       id = users[i].id;
-    //       profession = users[i].profession;
-    //     }
-    //   }
-    // }
-    // if(profession == "student"){
-    //   const events_students_con = await query('select * from events_students where event_id = ' + req.query.event_id + ' and student_id = ' + id +';');
-    //   if(events_students_con != 0){
-    //     bool = true;
-    //   }
-    // } else if (profession == "teacher"){
-    //   const events_teachers_con = await query('select * from events_teachers where event_id = ' + req.query.event_id + ' and teacher_id = ' + id +';');
-    //   if(events_teachers_con != 0){
-    //     bool = true;
-    //   }
-    // } else if (profession == "school"){
-    //   bool = true;
-    // }
-    // if(bool){
-    const events = await query('select * from events where id = ' + req.query.event_id + ';');
-    if(users.length != 0){
+    if (req.user.email) {
       for (i = 0; i < users.length; i++) {
-        // if(req.user.email == users[i].email){
-        //   sender = users[i];
-        // }
-        // if(req.query.receiver_id == users[i].id){
-        //   receiver = users[i];
-        // }
-        if(req.query.email == users[i].email){
-          sender = users[i];
-        }
-        if(req.query.receiver_id == users[i].id){
-          receiver = users[i];
+        if(req.user.email == users[i].email){
+          id = users[i].id;
+          profession = users[i].profession;
         }
       }
     }
-    eventlist = events;
-    obj = {
-      eventlist:eventlist,
-      sender:sender,
-      receiver:receiver
+    if(profession == "student"){
+      const events_students_con = await query('select * from events_students where event_id = ' + req.query.event_id + ' and student_id = ' + id +';');
+      if(events_students_con.length != 0){
+        bool = true;
+      }
+    } else if (profession == "teacher"){
+      const events_teachers_con = await query('select * from events_teachers where event_id = ' + req.query.event_id + ' and teacher_id = ' + id +';');
+      if(events_teachers_con.length != 0){
+        bool = true;
+      }
     }
-    res.json(obj);
-  // }
+    if(bool){
+      const events = await query('select * from events where id = ' + req.query.event_id + ';');
+      if(users.length != 0){
+        for (i = 0; i < users.length; i++) {
+          if(req.user.email == users[i].email){
+            sender = users[i];
+          }
+          if(req.query.receiver_id == users[i].id){
+            receiver = users[i];
+          }
+        }
+      }
+      eventlist = events;
+      obj = {
+        eventlist:eventlist,
+        sender:sender,
+        receiver:receiver,
+        status:200
+      }
+      res.json(obj);
+    } else{
+      console.log("マイページへ")
+      obj = {
+        status:400
+      }
+      res.json(obj);
+    }
   }
 
-  // if(req.user){
-    getESR();
-    // } else {
-    //     res.redirect("auth/signin")
-    // }
+  if(req.user){
+      getESR();
+  } else {
+    console.log("sign inへ")
+    obj = {
+      status:401
+    }
+    res.json(obj);
+  }
 });
 
 router.post('/', [body("action").not().isEmpty().withMessage("アクションを入力してください。").isNumeric().withMessage("アクションに数字を入力してください。"),
@@ -96,7 +99,7 @@ router.post('/', [body("action").not().isEmpty().withMessage("アクションを
                 　],(req, res, next) =>{
   const errors = validationResult(req);
   let profession;
-  let bool;
+  let bool = false;
   let obj;
   let sender;
   let receiver;
@@ -104,88 +107,79 @@ router.post('/', [body("action").not().isEmpty().withMessage("アクションを
   async function getESR(){
     const users = await query('select * from users;');
     const events = await query('select * from events where id = ' + req.body.event_id + ';');
-    // if (req.user.email) {
-    //   for (i = 0; i < users.length; i++) {
-    //     if(req.user.email == users[i].email){
-    //       id = users[i].id;
-    //       profession = users[i].profession;
-    //     }
-    //   }
-    // }
-    // if(profession == "student"){
-    //   const events_students_con = await query('select * from events_students where event_id = ' + req.body.event_id + ' and student_id = ' + id +';');
-    //   if(events_students_con != 0){
-    //     bool = true;
-    //   }
-    // } else if (profession == "teacher"){
-    //   const events_teachers_con = await query('select * from events_teachers where event_id = ' + req.body.event_id + ' and teacher_id = ' + id +';');
-    //   if(events_teachers_con != 0){
-    //     bool = true;
-    //   }
-    // } else if (profession == "school"){
-    //   bool = true;
-    // }
-    // if(bool){
-    if(users.length != 0){
+    if (req.user.email) {
       for (i = 0; i < users.length; i++) {
-        // if(req.user.email == users[i].email){
-        //   sender = users[i];
-        // }
-        // if(req.user.receiver_id == users[i].id){
-        //   receiver = users[i];
-        // }
-        if(req.body.email == users[i].email){
-          sender = users[i];
-        }
-        if(req.body.receiver_id == users[i].id){
-          receiver = users[i];
+        if(req.user.email == users[i].email){
+          id = users[i].id;
+          profession = users[i].profession;
         }
       }
     }
-    eventlist = events;
-    obj = {
-      eventlist:eventlist,
-      sender:sender,
-      receiver:receiver
+    if(profession == "student"){
+      const events_students_con = await query('select * from events_students where event_id = ' + req.body.event_id + ' and student_id = ' + id +';');
+      if(events_students_con.length != 0){
+        bool = true;
+      }
+    } else if (profession == "teacher"){
+      const events_teachers_con = await query('select * from events_teachers where event_id = ' + req.body.event_id + ' and teacher_id = ' + id +';');
+      if(events_teachers_con.length != 0){
+        bool = true;
+      }
     }
-    res.json(obj);
-  // }
+    if(bool){
+      if(users.length != 0){
+        for (i = 0; i < users.length; i++) {
+          if(req.user.email == users[i].email){
+            sender = users[i];
+          }
+          if(req.body.receiver_id == users[i].id){
+            receiver = users[i];
+          }
+        }
+      }
+      eventlist = events;
+      obj = {
+        eventlist:eventlist,
+        sender:sender,
+        receiver:receiver,
+        status:401
+      }
+      res.json(obj);
+    } else {
+      console.log("マイページへ")
+      obj = {
+        status:400
+      }
+      res.json(obj);
+    }
   }
   async function getSR(){
     const users = await query('select * from users;');
-    // if (req.user.email) {
-    //   for (i = 0; i < users.length; i++) {
-    //     if(req.user.email == users[i].email){
-    //       id = users[i].id;
-    //       profession = users[i].profession;
-    //     }
-    //   }
-    // }
-    // if(profession == "student"){
-    //   const events_students_con = await query('select * from events_students where event_id = ' + req.body.event_id + ' and student_id = ' + id +';');
-    //   if(events_students_con != 0){
-    //     bool = true;
-    //   }
-    // } else if (profession == "teacher"){
-    //   const events_teachers_con = await query('select * from events_teachers where event_id = ' + req.body.event_id + ' and teacher_id = ' + id +';');
-    //   if(events_teachers_con != 0){
-    //     bool = true;
-    //   }
-    // } else if (profession == "school"){
-    //   bool = true;
-    // }
+    if (req.user.email) {
+      for (i = 0; i < users.length; i++) {
+        if(req.user.email == users[i].email){
+          id = users[i].id;
+          profession = users[i].profession;
+        }
+      }
+    }
+    if(profession == "student"){
+      const events_students_con = await query('select * from events_students where event_id = ' + req.body.event_id + ' and student_id = ' + id +';');
+      if(events_students_con != 0){
+        bool = true;
+      }
+    } else if (profession == "teacher"){
+      const events_teachers_con = await query('select * from events_teachers where event_id = ' + req.body.event_id + ' and teacher_id = ' + id +';');
+      if(events_teachers_con != 0){
+        bool = true;
+      }
+    }
     if(users.length != 0){
       for (i = 0; i < users.length; i++) {
-        // if(req.user.email == users[i].email){
-        //   sender = users[i];
-        // }
-        // if(req.user.receiver_id == users[i].id){
-        //   receiver = users[i];
-        // }
-        if(req.body.email == users[i].email){
+        if(req.user.email == users[i].email){
           sender = users[i];
         }
-        if(req.body.receiver_id == users[i].id){
+        if(req.user.receiver_id == users[i].id){
           receiver = users[i];
         }
       }
@@ -195,31 +189,31 @@ router.post('/', [body("action").not().isEmpty().withMessage("アクションを
   let dtx;
   let handler;
   async function destroy(){
-    // if(bool){
-    address = req.body.event_id.toString() + "_" + sender.id.toString() + "_" + req.body.receiver_id.toString();
-    const evaluates = await query('select * from evaluates where event_id = ' + req.body.event_id.toString() + ' and sender_id = ' + sender.id.toString() + ' and receiver_id = ' + req.body.receiver_id.toString() + ';');
-    if(evaluates.length != 0){
-      dtx = iost.call("ContractCd8WSt1F3N8Kb7PF2JAEZM81HhxWqHKdKEDDvhpDvCDC", "destroy", [address]);
-      iost.signAndSend(dtx);
-      handler = iost.signAndSend(dtx);
-      handler.listen();
-      handler.onPending(console.log);
-      handler.onSuccess(console.log);
-      handler.onFailed(console.log);
-      connection.query('delete from evaluates where event_id = ' + req.body.event_id.toString() + ' and sender_id = ' + sender.id.toString() + ' and receiver_id = ' + req.body.receiver_id.toString() + ';', function(err, success){
-        if (err == null) {
-          connection.query('delete from evaluates_free where evaluate_id = ' + evaluates[0].id + ';', function(err, success2){
-            if (err == null) {
-            } else {
-              console.log(err);
-            }
-          });
-        } else {
-          console.log(err);
-        }
-      });
+    if(bool){
+      address = req.body.event_id.toString() + "_" + sender.id.toString() + "_" + req.body.receiver_id.toString();
+      const evaluates = await query('select * from evaluates where event_id = ' + req.body.event_id.toString() + ' and sender_id = ' + sender.id.toString() + ' and receiver_id = ' + req.body.receiver_id.toString() + ';');
+      if(evaluates.length != 0){
+        dtx = iost.call("ContractCd8WSt1F3N8Kb7PF2JAEZM81HhxWqHKdKEDDvhpDvCDC", "destroy", [address]);
+        iost.signAndSend(dtx);
+        handler = iost.signAndSend(dtx);
+        handler.listen();
+        handler.onPending(console.log);
+        handler.onSuccess(console.log);
+        handler.onFailed(console.log);
+        connection.query('delete from evaluates where event_id = ' + req.body.event_id.toString() + ' and sender_id = ' + sender.id.toString() + ' and receiver_id = ' + req.body.receiver_id.toString() + ';', function(err, success){
+          if (err == null) {
+            connection.query('delete from evaluates_free where evaluate_id = ' + evaluates[0].id + ';', function(err, success2){
+              if (err == null) {
+              } else {
+                console.log(err);
+              }
+            });
+          } else {
+            console.log(err);
+          }
+        });
+      }
     }
-  // }
   }
   let now = getStringFromDate(new Date());
   let freeStatement = "";
@@ -227,71 +221,80 @@ router.post('/', [body("action").not().isEmpty().withMessage("アクションを
   let hash;
   let atx;
   async function add(){
-    // if(bool){
-    if(req.body.free != null){
-      for(var i = 0; i<req.body.free.length; i++){
-        freeStatement = freeStatement + req.body.free[i].toString();
+    if(bool){
+      if(req.body.free != null){
+        for(var i = 0; i<req.body.free.length; i++){
+          freeStatement = freeStatement + req.body.free[i].toString();
+        }
       }
-    }
-    evaluationStatement = req.body.action.toString() + req.body.think.toString() + req.body.team.toString() + req.body.comments.toString() + address.toString() + freeStatement;
-    console.log(evaluationStatement);
-    hash = crypto.createHash('sha256').update(evaluationStatement, 'utf8').digest('hex');
-    atx = iost.call("ContractCd8WSt1F3N8Kb7PF2JAEZM81HhxWqHKdKEDDvhpDvCDC", "add", [address,hash]);
-    console.log(atx);
-    handler = iost.signAndSend(atx);
-    handler.listen();
-    handler.onPending(console.log);
-    handler.onSuccess(console.log);
-    handler.onFailed(console.log);
-    connection.query('insert into evaluates set ? ;', {
-      action: req.body.action,
-      think: req.body.think,
-      team: req.body.team,
-      comments: req.body.comments,
-      txhash: address,
-      event_id: req.body.event_id,
-      sender_id: sender.id,
-      receiver_id: req.body.receiver_id,
-      created_at: now,
-    },
-    function(err, success){
-      if (err == null) {
-        connection.query('select * from evaluates where txhash = "' + address + '";', function(err, evaluates2){
-          if(req.body.free != null){
-            for(var i = 0; i<req.body.free.length; i++){
-              connection.query('insert into evaluates_free set ? ;', {
-                evaluate_id: evaluates2[0].id,
-                free_id: req.body.free[i]
-              },function(err, success3){
-                if (err == null) {
-                } else {
-                  console.log(err);
+      evaluationStatement = req.body.action.toString() + req.body.think.toString() + req.body.team.toString() + req.body.comments.toString() + address.toString() + freeStatement;
+      console.log(evaluationStatement);
+      hash = crypto.createHash('sha256').update(evaluationStatement, 'utf8').digest('hex');
+      atx = iost.call("ContractCd8WSt1F3N8Kb7PF2JAEZM81HhxWqHKdKEDDvhpDvCDC", "add", [address,hash]);
+      console.log(atx);
+      handler = iost.signAndSend(atx);
+      handler.listen();
+      handler.onPending(console.log);
+      handler.onSuccess(console.log);
+      handler.onFailed(console.log);
+      connection.query('insert into evaluates set ? ;', {
+        action: req.body.action,
+        think: req.body.think,
+        team: req.body.team,
+        comments: req.body.comments,
+        txhash: address,
+        event_id: req.body.event_id,
+        sender_id: sender.id,
+        receiver_id: req.body.receiver_id,
+        created_at: now,
+      },
+      function(err, success){
+        if (err == null) {
+          connection.query('select * from evaluates where txhash = "' + address + '";', function(err, evaluates2){
+            if(req.body.free != null){
+              for(var i = 0; i<req.body.free.length; i++){
+                connection.query('insert into evaluates_free set ? ;', {
+                  evaluate_id: evaluates2[0].id,
+                  free_id: req.body.free[i]
+                },function(err, success3){
+                  if (err == null) {
+                  } else {
+                    console.log(err);
+                  }
                 }
+                );
               }
-              );
             }
-          }
-        });
-      } else {
-        console.log(err);
-      }
-    });
-  // }
+          });
+        } else {
+          console.log(err);
+        }
+      });
+    }
   }
   async function total(){
     await getSR();
     await destroy();
     await add();
+    console.log("参加者リストへ")
+    obj = {
+      status:200
+    }
+    res.json(obj);
   }
-  // if(req.user){
+  if(req.user){
     if (!errors.isEmpty()) {
       getESR();
     }else {
       total();
     }
-  // } else {
-  //     res.redirect("auth/signin")
-  // }
+  } else {
+    console.log("sign inへ")
+    obj = {
+      status:401
+    }
+    res.json(obj);
+  }
 });
 
 function getStringFromDate(date) {
